Break down component
